Add migration test cases for Bard API
