+++ .gitignore
@@ -1,3 +1,9 @@
+# Python files
+__pycache__/
+.pytest_cache
+*.pyc
+.env/
+
 # Binaries for programs and plugins
 *.exe
 *.exe~
